requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
nltk==3.9.1
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.0rc2
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
tqdm==4.67.1
Unidecode==1.4.0
urllib3==2.4.0
Werkzeug==3.1.3
